Refactor AB migration
